Add functions and keybinds for compiling a project and navigating to next/previous errors
I want to look over the next/prev error functions as they don't quite behave the way I like, additionally I want to look into compilation buffer hooks to customise what the compilation buffer looks like and how it behaves.

Note: evil navigation does not currently work in the compilation buffer, need to figure out why and fix it or workaround it somehow so that i can comfortably close the buffer and open other ones in that split if i want to.